requirements.txt
google-cloud-pubsub==2.13.0
google-cloud-storage==2.7.0
google-cloud-bigquery==3.5.0
tweepy==4.14.0
pandas==1.5.0
textblob==0.17.1
vaderSentiment==3.3.2
